refactor(structure): move some folders and files
