+++ src/Models/ModelGenerator.js
@@ -6,9 +6,10 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-import { BelongsToRelation } from '#src/Relations/BelongsToRelation'
-import { HasManyRelation } from '#src/Relations/HasManyRelation'
+
 import { HasOneRelation } from '#src/Relations/HasOneRelation'
+import { HasManyRelation } from '#src/Relations/HasManyRelation'
+import { BelongsToRelation } from '#src/Relations/BelongsToRelation'
 import { ManyToManyRelation } from '#src/Relations/ManyToManyRelation'
 
 export class ModelGenerator {
