+++ src/index.js
@@ -9,26 +9,26 @@
 
 import { Config } from '@athenna/config'
 
-import { QueryBuilder } from '#src/Builders/QueryBuilder'
 import { DriverFactory } from '#src/Factories/DriverFactory'
+import { QueryBuilder } from '#src/Database/Builders/QueryBuilder'
 
-export * from './Builders/Column.js'
-export * from './Builders/Criteria.js'
-export * from './Builders/ModelQueryBuilder.js'
-export * from './Builders/Relation.js'
-export * from './Builders/SchemaBuilder.js'
+export * from './Models/Column.js'
+export * from './Models/Criteria.js'
+export * from './Models/ModelQueryBuilder.js'
+export * from './Relations/Relation.js'
+export * from './Models/SchemaBuilder.js'
 export * from './Facades/Database.js'
 export * from './Facades/DB.js'
 export * from './Factories/ConnectionFactory.js'
 export * from './Factories/DriverFactory.js'
-export * from './Factories/ModelFactory.js'
-export * from './Generators/ModelGenerator.js'
+export * from './Models/ModelFactory.js'
+export * from './Models/ModelGenerator.js'
 export * from './Helpers/DatabaseLoader.js'
-export * from './Migrations/Migration.js'
+export * from './Database/Migrations/Migration.js'
 export * from './Models/Model.js'
-export * from './Resources/Resource.js'
-export * from './Seeders/Seeder.js'
-export * from './Transactions/Transaction.js'
+export * from './Models/Resource.js'
+export * from './Database/Seeders/Seeder.js'
+export * from './Database/Transactions/Transaction.js'
 
 export class DatabaseImpl {
   /**
