+++ tests/Stubs/app/Resources/UserResource.js
@@ -1,4 +1,4 @@
-import { Resource } from '#src/Resources/Resource'
+import { Resource } from '#src/Models/Resource'
 
 export class UserResource extends Resource {
   /**
